the py script file is the only one you need one you need to make it work on a picp pi the others show the diagrAM AND SIMULATOR files so you can open it on wokwi
maine.PY is the only file you need to run this on a Pico Pie. You just drag that into it and it goes. Or I should say after you've built the circuit. But if you wanna use this on WOW a WI online simulator then you can play with it and edit it and probably try to repurpose it and use it as a template. It represents about 80 hours of research trying to actually come out with these commands that don't involve all the structure the cumbersome structure that makes it impossible to understand. To me it's worth it to figure out how these things work in order to use them rather than just drop and drag somebody else's script in. This is completely editable and you and you can just use the subcommands if you want. Which I think is what makes my publicly accessible scripts invaluable do everyone who is interested in programming this thing The Pico Pie. or using micro python for that matter. Thank you for reading.